+++ src/metric/ir/dcg.cc
@@ -30,10 +30,10 @@ namespace ir {
 
 const std::string Dcg::NAME_ = "DCG";
 
-MetricScore Dcg::compute_dcg(const Label* labels, unsigned int len) const {
-  const unsigned int size = std::min(cutoff(), len);
+MetricScore Dcg::compute_dcg(const Label* labels, size_t len) const {
+  const size_t size = std::min(cutoff(), len);
   double dcg = 0.0;
-  for (unsigned int i = 0; i < size; ++i)
+  for (size_t i = 0; i < size; ++i)
     dcg += (pow(2.0, labels[i]) - 1.0f) / log2(i + 2.0f);
   return (MetricScore) dcg;
 }
@@ -56,12 +56,12 @@ MetricScore Dcg::evaluate_result_list(const quickrank::data::QueryResults* rl,
 
 std::unique_ptr<Jacobian> Dcg::jacobian(
     std::shared_ptr<data::RankedResults> ranked) const {
-  const unsigned int size = std::min(cutoff(), ranked->num_results());
+  const size_t size = std::min(cutoff(), ranked->num_results());
   std::unique_ptr<Jacobian> jacobian = std::unique_ptr<Jacobian>(
       new Jacobian(ranked->num_results()));
 
-  for (unsigned int i = 0; i < size; ++i) {
-    for (unsigned int j = i + 1; j < ranked->num_results(); ++j) {
+  for (size_t i = 0; i < size; ++i) {
+    for (size_t j = i + 1; j < ranked->num_results(); ++j) {
       // if the score is the same, non changes occur
       if (ranked->sorted_labels()[ranked->pos_of_rank(i)]
           != ranked->sorted_labels()[ranked->pos_of_rank(j)]) {
