+++ src/metric/ir/ndcg.cc
@@ -67,10 +67,10 @@ std::unique_ptr<Jacobian> Ndcg::jacobian(
   if (idcg <= 0.0)
     return jacobian;
 
-  const unsigned int size = std::min(cutoff(), ranked->num_results());
+  const size_t size = std::min(cutoff(), ranked->num_results());
 
-  for (unsigned int i = 0; i < size; ++i) {
-    for (unsigned int j = i + 1; j < ranked->num_results(); ++j) {
+  for (size_t i = 0; i < size; ++i) {
+    for (size_t j = i + 1; j < ranked->num_results(); ++j) {
       // if the score is the same, non changes occur
       if (ranked->sorted_labels()[i] != ranked->sorted_labels()[j]) {
         //*p_jacobian =
